luding breaking changes), run:   npm audit fix --force  Run `npm audit` for details. PS C:\Users\likho\OneDrive\GitHub\Deployo> open index.html open: The term 'open' is not recognized as a name of a cmdlet, function, script file, or executable program. Check the spelling of the name, or if a path was included, verify that the path is correct and try again.
luding breaking changes), run:
  npm audit fix --force

Run `npm audit` for details.
PS C:\Users\likho\OneDrive\GitHub\Deployo> open index.html
open: The term 'open' is not recognized as a name of a cmdlet, function, script file, or executable program.
Check the spelling of the name, or if a path was included, verify that the path is correct and try again.